fix(app): fix create a new instance when re-render a component
fix create a new instance when re-render a component
